Fixed issue where the VSIX installed extention didn't run.
This was caused because the dll included in the VSIX was not the one build by the SnowPlow project and didn't contain SnowPlow. The VSIX is now built by the same project as the dll.

Also, fixed issue where positive exit codes were interpreted as the unit tests didn't run, rather than just some failed.